+++ packages/realm_dart/lib/src/realm_class.dart
@@ -109,6 +109,8 @@ class Realm implements Finalizable {
   late final RealmMetadata _metadata;
   late final RealmHandle _handle;
   final bool _isInMigration;
+  late final RealmCallbackTokenHandle? _schemaCallbackHandle;
+  final List<StreamController<RealmSchemaChanges>> _schemaChangeListeners = [];
 
   /// An object encompassing this `Realm` instance's dynamic API.
   late final DynamicRealm dynamic = DynamicRealm._(this);
@@ -130,6 +132,18 @@ class Realm implements Finalizable {
 
   Realm._(this.config, [RealmHandle? handle, this._isInMigration = false]) : _handle = handle ?? _openRealm(config) {
     _populateMetadata();
+
+    // The schema of a Realm file may change due to sync adding new properties/classes. We subscribe for notifications
+    // in order to update the managed schema instance in case this happens. The same is true for dynamic Realms. For
+    // local Realms with user-supplied schema, the schema on disk may still change, but Core doesn't report the updated
+    // schema, so even if we subscribe, we wouldn't be able to see the updates.
+    if (config is FlexibleSyncConfiguration || config.schemaObjects.isEmpty) {
+      // TODO: enable once https://github.com/realm/realm-core/issues/7426 is fixed.
+      // _schemaCallbackHandle = realmCore.subscribeForSchemaNotifications(this);
+      _schemaCallbackHandle = null;
+    } else {
+      _schemaCallbackHandle = null;
+    }
   }
 
   /// A method for asynchronously opening a [Realm].
@@ -387,6 +401,7 @@ class Realm implements Finalizable {
       return;
     }
 
+    _schemaCallbackHandle?.release();
     realmCore.closeRealm(this);
     handle.release();
   }
@@ -596,6 +611,59 @@ class Realm implements Finalizable {
   Future<bool> refreshAsync() async {
     return realmCore.realmRefreshAsync(this);
   }
+
+  /// Allows listening for schema changes on this Realm. Only dynamic and synchronized
+  /// Realms will emit schema change notifications.
+  ///
+  /// Returns a [Stream] of [RealmSchemaChanges] that can be listened to.
+  // TODO: this is private due to https://github.com/realm/realm-core/issues/7426. Once that is fixed, we can expose it.
+  Stream<RealmSchemaChanges> get _schemaChanges {
+    late StreamController<RealmSchemaChanges> controller;
+    controller = StreamController<RealmSchemaChanges>(
+        onListen: () => _schemaChangeListeners.add(controller),
+        onPause: () => _schemaChangeListeners.remove(controller),
+        onResume: () => _schemaChangeListeners.add(controller),
+        onCancel: () => _schemaChangeListeners.remove(controller),
+        sync: true);
+    return controller.stream;
+  }
+
+  void _updateSchema() {
+    final newSchema = realmCore.readSchema(this);
+    for (final listener in _schemaChangeListeners) {
+      listener.add(RealmSchemaChanges._(schema, newSchema));
+    }
+
+    for (final schemaObject in newSchema) {
+      final existing = schema.firstWhereOrNull((s) => s.name == schemaObject.name);
+      if (existing == null) {
+        schema.add(schemaObject);
+        final meta = realmCore.getObjectMetadata(this, schemaObject);
+        metadata._add(meta);
+      } else if (schemaObject.length > existing.length) {
+        final existingMeta = metadata.getByName(schemaObject.name);
+        final propertyMeta = realmCore.getPropertiesMetadata(this, existingMeta.classKey, existingMeta.primaryKey);
+        for (final property in schemaObject) {
+          final existingProp = existing.firstWhereOrNull((e) => e.mapTo == property.mapTo);
+          if (existingProp == null) {
+            existing.add(property);
+            existingMeta[property.name] = propertyMeta[property.name]!;
+          }
+        }
+      }
+    }
+  }
+}
+
+/// Describes the schema changes that occurred on a Realm
+class RealmSchemaChanges {
+  /// The current schema, as returned by [Realm.schema].
+  final RealmSchema currentSchema;
+
+  /// The new schema that will be applied to the Realm as soon as the event handler completes.
+  final RealmSchema newSchema;
+
+  RealmSchemaChanges._(this.currentSchema, this.newSchema);
 }
 
 /// Provides a scope to safely write data to a [Realm]. Can be created using [Realm.beginWrite] or
@@ -790,6 +858,10 @@ extension RealmInternal on Realm {
   static void logMessageForTesting(Level logLevel, String message) {
     realmCore.logMessageForTesting(logLevel, message);
   }
+
+  void updateSchema() {
+    _updateSchema();
+  }
 }
 
 /// @nodoc
@@ -881,13 +953,19 @@ class RealmMetadata {
 
   RealmMetadata._(Iterable<RealmObjectMetadata> objectMetadatas) {
     for (final metadata in objectMetadatas) {
-      if (!metadata.schema.isGenericRealmObject) {
-        _typeMap[metadata.schema.type] = metadata;
-      } else {
-        _stringMap[metadata.schema.name] = metadata;
-      }
-      _classKeyMap[metadata.classKey] = metadata;
+      _add(metadata);
     }
+  }
+
+  /// This is used when constructing the metadata, but also when the Realm schema
+  /// changes.
+  void _add(RealmObjectMetadata metadata) {
+    if (!metadata.schema.isGenericRealmObject) {
+      _typeMap[metadata.schema.type] = metadata;
+    } else {
+      _stringMap[metadata.schema.name] = metadata;
+    }
+    _classKeyMap[metadata.classKey] = metadata;
   }
 
   RealmObjectMetadata getByType(Type type) {