+++ packages/realm_dart/test/geospatial_test.realm.dart
@@ -69,12 +69,15 @@ class Location extends _Location
   static final schema = () {
     RealmObjectBase.registerFactory(Location._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(ObjectType.embeddedObject, Location, 'Location', [
+    return SchemaObject(ObjectType.embeddedObject, Location, 'Location', [
       SchemaProperty('type', RealmPropertyType.string),
       SchemaProperty('coordinates', RealmPropertyType.double,
           collectionType: RealmCollectionType.list),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class Restaurant extends _Restaurant
@@ -133,13 +136,15 @@ class Restaurant extends _Restaurant
   static final schema = () {
     RealmObjectBase.registerFactory(Restaurant._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(
-        ObjectType.realmObject, Restaurant, 'Restaurant', [
+    return SchemaObject(ObjectType.realmObject, Restaurant, 'Restaurant', [
       SchemaProperty('name', RealmPropertyType.string, primaryKey: true),
       SchemaProperty('location', RealmPropertyType.object,
           optional: true, linkTarget: 'Location'),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class LocationList extends _LocationList
@@ -189,10 +194,12 @@ class LocationList extends _LocationList
   static final schema = () {
     RealmObjectBase.registerFactory(LocationList._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(
-        ObjectType.realmObject, LocationList, 'LocationList', [
+    return SchemaObject(ObjectType.realmObject, LocationList, 'LocationList', [
       SchemaProperty('locations', RealmPropertyType.object,
           linkTarget: 'Location', collectionType: RealmCollectionType.list),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }