+++ packages/realm_dart/test/realm_map_test.realm.dart
@@ -60,11 +60,14 @@ class Car extends _Car with RealmEntity, RealmObjectBase, RealmObject {
   static final schema = () {
     RealmObjectBase.registerFactory(Car._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(ObjectType.realmObject, Car, 'Car', [
+    return SchemaObject(ObjectType.realmObject, Car, 'Car', [
       SchemaProperty('make', RealmPropertyType.string, primaryKey: true),
       SchemaProperty('color', RealmPropertyType.string, optional: true),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class EmbeddedValue extends _EmbeddedValue
@@ -111,11 +114,14 @@ class EmbeddedValue extends _EmbeddedValue
   static final schema = () {
     RealmObjectBase.registerFactory(EmbeddedValue._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(
+    return SchemaObject(
         ObjectType.embeddedObject, EmbeddedValue, 'EmbeddedValue', [
       SchemaProperty('intValue', RealmPropertyType.int),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class TestRealmMaps extends _TestRealmMaps
@@ -441,7 +447,7 @@ class TestRealmMaps extends _TestRealmMaps
   static final schema = () {
     RealmObjectBase.registerFactory(TestRealmMaps._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(
+    return SchemaObject(
         ObjectType.realmObject, TestRealmMaps, 'TestRealmMaps', [
       SchemaProperty('key', RealmPropertyType.int, primaryKey: true),
       SchemaProperty('boolMap', RealmPropertyType.bool,
@@ -492,4 +498,7 @@ class TestRealmMaps extends _TestRealmMaps
           optional: true, collectionType: RealmCollectionType.map),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }