+++ packages/realm_dart/test/realm_value_test.realm.dart
@@ -58,10 +58,13 @@ class TuckedIn extends _TuckedIn
   static final schema = () {
     RealmObjectBase.registerFactory(TuckedIn._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(ObjectType.embeddedObject, TuckedIn, 'TuckedIn', [
+    return SchemaObject(ObjectType.embeddedObject, TuckedIn, 'TuckedIn', [
       SchemaProperty('x', RealmPropertyType.int),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class AnythingGoes extends _AnythingGoes
@@ -149,8 +152,7 @@ class AnythingGoes extends _AnythingGoes
   static final schema = () {
     RealmObjectBase.registerFactory(AnythingGoes._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(
-        ObjectType.realmObject, AnythingGoes, 'AnythingGoes', [
+    return SchemaObject(ObjectType.realmObject, AnythingGoes, 'AnythingGoes', [
       SchemaProperty('oneAny', RealmPropertyType.mixed,
           optional: true, indexType: RealmIndexType.regular),
       SchemaProperty('manyAny', RealmPropertyType.mixed,
@@ -161,6 +163,9 @@ class AnythingGoes extends _AnythingGoes
           optional: true, collectionType: RealmCollectionType.set),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class Stuff extends _Stuff with RealmEntity, RealmObjectBase, RealmObject {
@@ -213,8 +218,11 @@ class Stuff extends _Stuff with RealmEntity, RealmObjectBase, RealmObject {
   static final schema = () {
     RealmObjectBase.registerFactory(Stuff._);
     register(_toEJson, _fromEJson);
-    return const SchemaObject(ObjectType.realmObject, Stuff, 'Stuff', [
+    return SchemaObject(ObjectType.realmObject, Stuff, 'Stuff', [
       SchemaProperty('i', RealmPropertyType.int),
     ]);
   }();
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }